+++ src/main/kotlin/de/thm/mni/ii/classroom/model/classroom/Ticket.kt
@@ -1,4 +1,4 @@
-package de.thm.mni.ii.classroom.model
+package de.thm.mni.ii.classroom.model.classroom
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
