+++ src/main/kotlin/de/thm/mni/ii/classroom/model/classroom/UserRole.kt
@@ -1,4 +1,4 @@
-package de.thm.mni.ii.classroom.model
+package de.thm.mni.ii.classroom.model.classroom
 
 import org.springframework.security.core.GrantedAuthority
