+++ src/main/kotlin/de/thm/mni/ii/classroom/model/message/Message.kt
@@ -1,0 +1,9 @@
+package de.thm.mni.ii.classroom.model.message
+
+import java.time.LocalDateTime
+
+data class Message(
+    val jwt: String,
+    val timestamp: LocalDateTime,
+    val message: String
+)
